media_analysis/requirements.txt
anyio                     4.2.0
appnope                   0.1.3
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 2.4.1
async-lru                 2.0.4
attrs                     23.2.0
Babel                     2.14.0
beautifulsoup4            4.12.3
bleach                    6.1.0
build                     1.2.2.post1
certifi                   2024.7.4
cffi                      1.16.0
charset-normalizer        3.1.0
comm                      0.2.1
contourpy                 1.2.1
cycler                    0.12.1
debugpy                   1.8.0
decorator                 5.1.1
defusedxml                0.7.1
et-xmlfile                1.1.0
exceptiongroup            1.2.0
executing                 2.0.1
fastjsonschema            2.19.1
filelock                  3.18.0
fonttools                 4.50.0
fqdn                      1.5.1
fsspec                    2025.3.0
h11                       0.14.0
httpcore                  1.0.5
httpx                     0.27.0
huggingface-hub           0.29.3
idna                      3.4
importlib-metadata        7.0.1
importlib_resources       6.4.0
iniconfig                 2.0.0
ipykernel                 6.29.0
ipympl                    0.9.4
ipython                   8.18.1
ipython-genutils          0.2.0
ipywidgets                8.1.3
isoduration               20.11.0
jedi                      0.19.1
Jinja2                    3.1.3
joblib                    1.4.2
json5                     0.9.14
jsonpointer               2.4
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
jupyter_client            8.6.0
jupyter_core              5.7.1
jupyter-events            0.9.0
jupyter-lsp               2.2.2
jupyter_server            2.12.5
jupyter_server_terminals  0.5.2
jupyterlab                4.2.3
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.2
jupyterlab_widgets        3.0.11
kiwisolver                1.4.5
lxml                      5.3.0
MarkupSafe                2.1.4
matplotlib                3.8.4
matplotlib-inline         0.1.6
media-analysis            1.0.0           /Users/macbookpro/Desktop/hackathon/media_analysis
mistune                   3.0.2
mpmath                    1.3.0
nbclient                  0.9.0
nbconvert                 7.14.2
nbformat                  5.9.2
nest-asyncio              1.6.0
networkx                  3.2.1
notebook_shim             0.2.3
numpy                     1.26.4
opencv-python             4.6.0.66
openpyxl                  3.1.2
overrides                 7.6.0
packaging                 23.2
pandas                    2.2.1
pandocfilters             1.5.1
parso                     0.8.3
patsy                     0.5.6
pexpect                   4.9.0
pillow                    10.3.0
pip                       25.0.1
platformdirs              4.1.0
plotly                    5.23.0
pluggy                    1.5.0
prometheus-client         0.19.0
prompt-toolkit            3.0.43
psutil                    5.9.8
ptyprocess                0.7.0
pure-eval                 0.2.2
py-cpuinfo                9.0.0
pycparser                 2.21
Pygments                  2.17.2
pyparsing                 3.1.2
pyproject_hooks           1.2.0
pytest                    8.3.5
python-dateutil           2.8.2
python-json-logger        2.0.7
pytz                      2024.1
PyYAML                    6.0.1
pyzmq                     25.1.2
referencing               0.32.1
regex                     2024.11.6
requests                  2.31.0
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.17.1
safetensors               0.5.3
scikit-learn              1.5.1
scipy                     1.13.1
seaborn                   0.13.2
Send2Trash                1.8.2
setuptools                76.1.0
simpy                     4.1.1
six                       1.16.0
sniffio                   1.3.0
soupsieve                 2.5
stack-data                0.6.3
statsmodels               0.14.2
sympy                     1.13.3
tenacity                  9.0.0
terminado                 0.18.0
threadpoolctl             3.5.0
tinycss2                  1.2.1
tokenizers                0.21.1
tomli                     2.0.1
torch                     2.2.2
torchvision               0.17.2
tornado                   6.4
tqdm                      4.67.1
traitlets                 5.14.1
transformers              4.49.0
types-python-dateutil     2.8.19.20240106
typing_extensions         4.9.0
tzdata                    2024.1
ultralytics-thop          2.0.14
uri-template              1.3.0
urllib3                   2.0.3
wcwidth                   0.2.13
webcolors                 1.13
webencodings              0.5.1
websocket-client          1.7.0
wheel                     0.45.1
widgetsnbextension        4.0.11
zipp                      3.17.0